Scale key to 20x20
I decided to change the color of the key. I think yellow looks a bit better than green. This means I'll have to change the color of the lock too.

+++ Buttons/Game/key component.hpp
@@ -12,7 +12,7 @@
 #include <cstdint>
 
 struct Key {
-  uint16_t index;
+  uint8_t index;
 };
 
 #endif
